+++ src/api/authentication/authenticateRoutes.js
@@ -4,7 +4,7 @@ import jwt from 'jsonwebtoken';
 import koa from 'koa';
 import koaRoute from 'koa-route';
 import methodFilter from '../lib/middlewares/methodFilter';
-import rateLimiter from './rateLimiter';
+import rateLimiter from '../lib/rateLimiter';
 import userRepositoryFactory from '../users/userModel';
 
 const app = koa();
